+++ packages/nimble-components/src/rich-text-mention/users/view/styles.ts
@@ -1,9 +1,11 @@
 import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
 import {
-    mentionFont,
     mentionDisabledFontColor,
-    bodyFontColor
+    bodyFontColor,
+    bodyFont,
+    mentionFont,
+    bodyDisabledFontColor
 } from '../../../theme-provider/design-tokens';
 
 export const styles = css`
@@ -11,12 +13,13 @@ export const styles = css`
 
     :host {
         box-sizing: border-box;
-        font: ${mentionFont};
+        font: ${bodyFont};
         color: ${bodyFontColor};
         white-space: normal;
     }
 
     .control {
+        font: ${mentionFont};
         color: var(--ni-nimble-private-mention-font-color);
         display: none;
     }
@@ -27,6 +30,10 @@ export const styles = css`
 
     :host([disabled]) .control {
         color: ${mentionDisabledFontColor};
+    }
+
+    :host([disabled]) {
+        color: ${bodyDisabledFontColor};
     }
 
     :host([disable-editing]) slot {